pyproject.toml
[project]
name = "bftp"
version = "0.1.0"
description = ""
authors = [
    {name = "Joe McCain III",email = "jo3mccain@icloud.com"}
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "tensorflow (>=2.19.0,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "openai (>=1.76.0,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "tortoise-orm (>=0.25.0,<0.26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
